refactor: organize economy code
