+++ bot/handler/ecom/account.js
@@ -1,5 +1,5 @@
-import * as cache from '../storage/redis.js';
-import * as db from '../storage/postgres.js';
+import * as cache from '../../../storage/redis.js';
+import * as db from '../../../storage/postgres.js';
 
 const accounts = new Map();
